Reorganize refs tests into dedicated folder
Group refs-specific coverage under SIL.Harmony.Tests/Refs so the refs surface and behavior tests are easier to browse without mixing them into the core test root.
