+++ src/SIL.Harmony.Tests/Refs/TagCheckoutTests.cs
@@ -1,9 +1,9 @@
+using Microsoft.Extensions.DependencyInjection;
 using SIL.Harmony.Refs;
 using SIL.Harmony.Sample.Models;
-using Microsoft.Extensions.DependencyInjection;
 using RefTag = SIL.Harmony.Refs.Entities.Tag;
 
-namespace SIL.Harmony.Tests;
+namespace SIL.Harmony.Tests.Refs;
 
 public class TagCheckoutTests : DataModelTestBase
 {
